Вносит требуемые правки по ревью ПР
