+++ js/api.js
@@ -1,89 +1,5 @@
-import {showAlert, isEscapeKey} from './util.js';
-
-const body = document.querySelector('body');
-const templateSuccess = document.querySelector('#success').content.querySelector('.success');
-const templateError = document.querySelector('#error').content.querySelector('.error');
-const successUnit = templateSuccess.cloneNode(true);
-const successButton = successUnit.querySelector('.success__button');
-const errorUnit = templateError.cloneNode(true);
-const errorButton = errorUnit.querySelector('.error__button');
-
-const onSuccessMessageEscKeyDown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    successUnit.classList.add('hidden');
-  }
-};
-
-const onErrorMessageEscKeyDown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    errorUnit.classList.add('hidden');
-  }
-};
-
-const createSuccessMessage = () => {
-  body.appendChild(successUnit);
-  successUnit.classList.add('hidden');
-};
-
-const createErrorMessage = () => {
-  body.appendChild(errorUnit);
-  errorUnit.classList.add('hidden');
-};
-
-createSuccessMessage();
-createErrorMessage();
-
-function openSuccessMessage () {
-  successUnit.classList.remove('hidden');
-  document.addEventListener('keydown', onSuccessMessageEscKeyDown);
-
-  successButton.addEventListener('click', () => {
-    closeSuccessMessage ();
-  });
-
-  window.addEventListener('click', (event) => {
-    event.stopPropagation();
-    if (event.target.contains(successUnit)) {
-      successUnit.classList.add('hidden');
-    }
-  });
-}
-
-function closeSuccessMessage () {
-  successUnit.classList.add('hidden');
-
-  document.removeEventListener('keydown', onSuccessMessageEscKeyDown);
-
-  successButton.removeEventListener('click', () => {
-    closeSuccessMessage ();
-  });
-}
-
-function openErrorMessage () {
-  errorUnit.classList.remove('hidden');
-  document.addEventListener('keydown', onErrorMessageEscKeyDown);
-
-  errorButton.addEventListener('click', () => {
-    closeErrorMessage ();
-  });
-
-  window.addEventListener('click', (event) => {
-    event.stopPropagation();
-    if (event.target.contains(errorUnit)) {
-      errorUnit.classList.add('hidden');
-    }
-  });
-}
-
-function closeErrorMessage () {
-  errorUnit.classList.add('hidden');
-  document.removeEventListener('keydown', onErrorMessageEscKeyDown);
-  errorButton.removeEventListener('click', () => {
-    closeErrorMessage ();
-  });
-}
+import {showAlert} from './util.js';
+import {openErrorMessage, openSuccessMessage} from './messages.js';
 
 const getData = (onSuccess) => {
   fetch('https://25.javascript.pages.academy/kekstagram/data')
