--- queries.http
@@ -1,20 +1,0 @@
-HEAD /posts HTTP/1.1
-Host: jsonplaceholder.typicode.com
-
-###
-
-GET /posts HTTP/1.1
-Host: jsonplaceholder.typicode.com
-
-###
-
-POST /posts HTTP/1.1
-Host: jsonplaceholder.typicode.com
-Content-Type: application/json
-Content-Length: 88
-
-{
-    "title": "Новая публикация",
-    "body": "Текст публикации",
-    "userId": 31337
-}
